refactor: split chat from agent, drop client connect, naming convergence
One-shot cutover of the first-tree-hub CLI surface. Zero deprecation alias,
all references swept in lockstep (hub repo + first-tree-context).

## Top-level structure
- New `chat` command group: `list / history / send / open`
  (replaces `agent send / chats / history / chat`)
- Drop `client connect <url>` entirely (incl. username/password auth path);
  top-level `first-tree-hub connect <token>` is the sole onboarding entry point
- Drop top-level `config` command group; capability moves to `client config
  show / set / get` on the local `client.yaml`

## Renames
- `server db:migrate` → `server migrate`
- `server admin:create` → `server admin create` (subcommand group)
- `client hub-list` / `hub-disconnect` → `client list` / `disconnect`
- `agent sessions <name>` → `agent session list <name>` (joins existing
  `session suspend / terminate` group)
- `agent config get` → `agent config show`
- `agent register` / `agent pull` → `agent debug register` / `agent debug pull`
  (commander `{ hidden: true }`; hidden from `agent --help`)

## Cross-cutting sweeps
- CLAUDE.md injection template (`client/src/runtime/bootstrap.ts:generateToolsDoc`)
  now writes `first-tree-hub chat send`; every agent shelled out by the runtime
  will see the new command in its system prompt
- Server-side user-facing error hints in `services/message.ts` updated so agents
  that hit the unknown-command edge get pointed at `first-tree-hub chat send`
- Web UI Quick Start references updated
- skill (`skills/first-tree-hub-cli/`) and `evals.json` rewritten end-to-end

## Out of scope (per design doc §7)
Top-level `status` / `doctor` aggregation, daemon, removal of the `server`
command group, `org bind-tree` rename, and the Dockerfile CMD are deferred to
the broader first-tree CLI consolidation (first-tree-context PR 214).

## Version
`@agent-team-foundation/first-tree-hub` 0.12.10 → 0.13.0 (minor — breaking
changes to the published CLI surface).

## Cross-project dependency
Requires lockstep merge with first-tree-context PR #267
(docs sweep of `agent-hub/*.md`).
