+++ AGENTS.md
@@ -73,7 +73,7 @@ Full guide (rules, parallel dev installs, what's NOT isolated, teardown): [docs/
 
 **PostgreSQL only:** No Redis / MQ. PG covers storage, queuing (SKIP LOCKED), and notifications (LISTEN/NOTIFY).
 
-**Unified user-JWT auth:** Single user JWT (issued by `client connect`, stored at `~/.first-tree/hub/config/credentials.json`) authorizes both Web/Admin API and every agent the user manages on the Client WebSocket. JWT shape, route classification, and middleware choice live in [docs/http-path-conventions.md](docs/http-path-conventions.md) — this section covers only the runtime *binding* facts not in that spec. Agents bind via `agents.client_id` + a server-pushed `agent:pinned` frame; **R-RUN** is re-evaluated at every `agent:bind` against the live `agents → manager → user` join (cross-org under one user is allowed; revoked memberships refuse the bind immediately). Switching user requires `first-tree-hub client claim --confirm`, which atomically transfers `clients.user_id` and unpins the previous owner's agents. Per-agent `aghub_*` tokens retired by migration `0020`. Background: [docs/decouple-client-from-identity-design-zh.md](docs/decouple-client-from-identity-design-zh.md) (client-from-identity decoupling) and [proposals/hub-strip-jwt-ambient-scope.20260508.md](../first-tree-context/proposals/hub-strip-jwt-ambient-scope.20260508.md) (JWT-scope strip + route refactor).
+**Unified user-JWT auth:** Single user JWT (issued by `first-tree-hub connect <token>`, stored at `~/.first-tree/hub/config/credentials.json`) authorizes both Web/Admin API and every agent the user manages on the Client WebSocket. JWT shape, route classification, and middleware choice live in [docs/http-path-conventions.md](docs/http-path-conventions.md) — this section covers only the runtime *binding* facts not in that spec. Agents bind via `agents.client_id` + a server-pushed `agent:pinned` frame; **R-RUN** is re-evaluated at every `agent:bind` against the live `agents → manager → user` join (cross-org under one user is allowed; revoked memberships refuse the bind immediately). Switching user requires `first-tree-hub client claim --confirm`, which atomically transfers `clients.user_id` and unpins the previous owner's agents. Per-agent `aghub_*` tokens retired by migration `0020`. Background: [docs/decouple-client-from-identity-design-zh.md](docs/decouple-client-from-identity-design-zh.md) (client-from-identity decoupling) and [proposals/hub-strip-jwt-ambient-scope.20260508.md](../first-tree-context/proposals/hub-strip-jwt-ambient-scope.20260508.md) (JWT-scope strip + route refactor).
 
 **Inbox is the Server/Client boundary:** Server writes to Inbox (fan-out on write), Client pulls / receives WebSocket notifications. At-least-once delivery; Client is responsible for deduplication.
