+++ docs/agent-naming-design.md
@@ -313,7 +313,7 @@ name := lowercase(input)
 
 - **Agent name 保留字黑名单是否公开？** 显示在创建表单帮助文案会暴露系统结构；或只在碰撞时才告知。倾向：碰撞时告知 + 在文档页列出完整名单
 - **Mention 自动补全的匹配策略**：display name 是否做内部字符模糊匹配？agent name 是否只做前缀匹配？倾向：display name 做模糊（体验好），agent name 做前缀（精准）
-- **CLI：`first-tree-hub agent send @coder-agent` 是推荐写法，还是 `@` 可选？** 倾向：两者都接受，内部统一剥离 `@` 前缀
+- **CLI：`first-tree-hub chat send @coder-agent` 是推荐写法，还是 `@` 可选？** 倾向：两者都接受，内部统一剥离 `@` 前缀
 
 > 原草案中"API 响应是否除了 `name` 也暴露 `handle` 字段"这一问题已**作废** —— 因为 UI 术语已统一为 "Agent name"（不引入 "handle" 新概念），API 字段继续用 `name` 即可，下游无迁移压力
