+++ packages/command/src/commands/onboard.ts
@@ -17,10 +17,10 @@ async function promptMissing(args: Record<string, unknown>): Promise<void> {
     }
   }
 
-  // 2. Require that the user has already run `first-tree-hub client connect`
+  // 2. Require that the user has already run `first-tree-hub connect <token>`
   const { loadCredentials } = await import("../core/bootstrap.js");
   if (!loadCredentials()) {
-    throw new Error("No saved credentials. Run `first-tree-hub client connect <server-url>` before onboarding.");
+    throw new Error("No saved credentials. Run `first-tree-hub connect <token>` before onboarding.");
   }
 
   if (!args.id) {
