+++ packages/command/src/core/index.ts
@@ -84,7 +84,7 @@ export {
   uninstallClientService,
 } from "./service-install.js";
 export type { ExecuteUpdateResult, InstallMode } from "./update.js";
-// Self-update glue — exported so both `client start` and `client connect`
+// Self-update glue — exported so both `client start` and `connect <token>`
 // can pass identical prompt / install callbacks to the ClientRuntime.
 export { detectInstallMode, fetchLatestVersion, installGlobalLatest, PACKAGE_NAME } from "./update.js";
 export { createExecuteUpdate, declineUpdate, promptUpdate, SELF_RESTART_EXIT_CODE } from "./update-glue.js";
