+++ packages/server/src/api/agents.ts
@@ -176,7 +176,7 @@ export async function agentRoutes(app: FastifyInstance): Promise<void> {
     if (health === "disconnected") {
       return reply.status(200).send({
         status: "offline",
-        message: "Agent is not connected. Start the client with: first-tree-hub client connect <server-url>",
+        message: "Agent is not connected. Connect the client with: first-tree-hub connect <token>",
         connection,
       });
     }
