+++ packages/server/src/services/agent.ts
@@ -183,7 +183,7 @@ async function resolveAgentClient(
   if (!client.userId) {
     throw new BadRequestError(
       `Client "${data.clientId}" has not been claimed by a user yet. Have the operator run ` +
-        "`first-tree-hub client connect` on that machine before pinning an agent to it.",
+        "`first-tree-hub connect <token>` on that machine before pinning an agent to it.",
     );
   }
   if (client.userId !== manager.userId) {
